requirements.txt
awesome-slugify==1.6.5
hdx-python-api==2.3.6
